pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "a2a-wrapper"
version = "0.2.0"
description = "Production-friendly helpers for exposing agents as A2A servers and consuming them through the official A2A SDK."
readme = "README.md"
requires-python = ">=3.11"
keywords = ["a2a", "agent", "sdk", "langchain", "crewai", "wrapper"]
authors = [
  { name = "Ayush" }
]
license = { text = "MIT" }
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "a2a-sdk[http-server]",
  "httpx>=0.27",
  "uvicorn>=0.30",
]

[project.optional-dependencies]
langchain = [
  "langchain",
  "langchain-community",
  "langchain-openai",
  "langgraph",
  "python-dotenv",
]
crewai = [
  "crewai",
  "crewai-tools",
  "python-dotenv",
]
dev = [
  "build>=1.2",
  "pytest>=8.0",
  "twine>=5.0",
]

[project.urls]
Homepage = "https://github.com/your-org/a2a-wrapper"
Documentation = "https://github.com/your-org/a2a-wrapper#readme"
Repository = "https://github.com/your-org/a2a-wrapper"

[project.scripts]
a2a-wrapper-client = "a2a_wrapper._client:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["a2a_wrapper*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"
